+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# Changelog version 3.0.10
+
+## Added
+
+- Added compatibility with patch 11.1.5.
+- Added 665 icons and 96 musics from patch 11.1.0 and 11.1.5.
+  - Added 13 icons for Cataclysm Classic.
+  - Added 4 icons and 3 musics for Classic Era.
+
+## Fixed
+
+- Fixed "Open profile" not showing when right-clicking a player from a community member list.
+- Fixed an issue when using Prat where the setting to remove realm from character names wasn't working.
+
+## Removed
+
+- Removed the maximize button.
+- Removed the Wago Analytics module.
+
 # Changelog version 3.0.9
 
 ## Fixed